Write the new slice/encoder.
o) Track declarations needed by a slice and its children.
o) Look for redundant data using algorithms explained in the code.
o) Get rid of trying to slice the data up into smaller working units.  The
   only gains that gives are related to locality of reference, which is
   absolutely broken now, but likely to be fixed.

Performance is below that of the old encoder by a few times in the case of
data with a lot of redundancy, but it's probably fairly easy to fix that,
at least for some data sets.  Comments explain the plan-of-action and the
perceived problem space.

Random data performs about as well, it seems, in terms of streaming rate.

And even if performance is worse, that's to be expected since there's no
more crazy inlining, and probably a worthwhile trade-off since this version
of the encoder is humanly-comprehensible (and full of almost-useful
comments.)