Rewrite the XCodecEncoder encode function... o) Do the entirety of processing in the first loop rather than processing    references and then declarations.  This improves handling of in-stream    collisions and references to data earlier in a stream.    XXX This duplicates a lot of code.  Add a TODO item to clean it up. o) Correct an implementation bug: when we have a collision, we should still    check to see whether the last-known-good hash should be declared. o) As a result of merging the loops add logic to handle adjacent hash collisions    (and hits) gracefully.
The net effect of this is that performance may be slightly degraded in terms of
encode throughput (merging the loops was at first a boost in throughput but
handling edge cases and better encoding possibilities ate the increase) the
compression ratio for non-random, non-uniform data is significantly better than
previously, from 1% to 10% better in some tested data sets.  Random data should
be the same and veryuniform data sees a significant performance boost from the
elimination of the second loop (both in terms of eliminating the number of loops
and in terms of registering declarations earlier) -- about twice the throughput
of the previous encoder.