Be more strict about zero-length operations and add some nearby assertions.
One pressing example:

If we don't allow Buffer::skip(n) for any n when the Buffer is empty, and we
don't allow Buffer::skip(n) on any Buffer with less than n bytes, then we
cannot allow Buffer::skip(0) on any Buffer.  This has implications for the
likes of Buffer::moveout(), too.