Fix a long-standing mistake:     Channel::close should not take an EventCallback.
All the close routine does is let the caller know when it can
delete the channel in question.  There is nothing reasonable a
caller can do in the case of an error.

Also:
o) Make TypedCallback check whether it has been given a parameter,
   and require such.
o) Fix socket-accept-connect1 which had accidentally been
   disconnected from the build.