Add a moveout(Buffer *) overload to move contents from one Buffer to another.
This accomplishes the common idiom of:
	output_.append(input);
	input->clear();

But because of the semantic clarity of the function, it's possible to do some
optimization, namely to avoid bumping and dropping references gratuitously.  In
doing so, we achieve modest performance gains in macrobenchmarks.  The added
clarity in reading the code is nice, too.

While here also remove some OpenBSD-related cruft.