Fix deadline calculation.
We set the deadline in terms of the number of milliseconds away the
deadline is and then add the current time.  The previous approach would
do the arithmetic itself on the fields of NanoTime and got it wrong, so
that increments not modular to 1s would generate an invalid NanoTime,
which could cause things like pthread_cond_timedwait to fail.