Resolve locking issues in SocketUinet.
SocketUinet has had some historical issues (as do other codebases
using libuinet callbacks) with lock order, which this commit
resolves by using a special mutex to defer work from upcalls if
the locks cannot be acquired safely.

This code, while an improvement, may still have issues, but at
least seems to resolve the most common deadlocks and races which
were present in SocketUinet previously.

Other work in WANProxy may make some of this redundant in time,
particularly if callbacks and actions become more permanent, if
locking is added to callbacks (with interlock on dispatch),
etc.