Initial planning for more multithreading.
This provides an API for callbacks to be scheduled into a pool of
workers rather than just onto EventThread.  This is opt-in, and keeps
the behaviour of EventThread being single-threaded.

Note that PipeProducer is a logical place to hook in a new scheduler.

Add a scheduler field to Splice, which is a very logical unit for
thread splitting.  Don't use it yet.  Need to think hard about
whether we want to set this at construct time, or to allow pinning
afterwards instead.  Don't want to do this both ways.