Eliminate argument-capture callbacks.
These served to carry state to shared callback methods, but in
practice these were not used widely, and where they did they
carried a kind of state that was really not optimal.  Instead,
indicate things like which socket has been closed by having a
different method for each.  These methods are easier to
understand and clearer in purpose.