Use persistent cancellations for CallbackThread.
This reverses the cancellation so that callbacks track whether they
have been scheduled, and cancel themselves with the scheduler they
are attached to.

This also pins callbacks to a scheduler after their first schedule.

This results in something like a 10% reduction in overhead on
callbacks and scheduling.  Further improvements are expected once
callbacks themselves persist.