Nest the Method instance in SimpleCallback.
This is not really necessary for SimpleCallback, but is for consistency
with what the other callback types will go through.  This is to work
around the lack of typedef templates (and trying to avoid C++11-style
using declarations for now) for implementing things like EventCallback,
without exposing implementation details to consumers.