Checkpoint persistent callbacks II.
This converts EventCallback callbacks to be persistent callbacks,
as was done for SimpleCallback.

The only interesting difference over SimpleCallback is that we have
to be careful to steal the parameter to a TypedCallback and reset
it before executing the callback, or we can't reschedule the callback
from within its execution, which is a common pattern when two classes
are moving data between each other.