Get rid of the Buffer in Event.
This creates a new BufferEventCallback for the few kinds of events
that actually move out data.  Most kinds of events and over half of
the events we execute in practice don't move data with the event.
Only reads and their ilk do.

This also highlights the need to pass a reference to the Buffer,
not a copy of it, so that our refcount is low, and we can modify
the BufferSegments rather than keeping around a pristine copy on
the stack in the callback execution path.  Ugh.