Simplify CallbackThread scheduling.
We make 'select' a method of CallbackThread, rather than doing all the
work in the main body, and have it try everything in the list looking
for a viable interlock.  This avoids modifying the list, and makes the
logic simpler overall, as well as making dropping in other scheduling
algorithms much simpler.