Fix bug #18; adjust for OpenSSL 1.1.0+ opaque structures.
OpenSSL has finally decided that struct fields aren't a very good API, and
there are new APIs to replace code that worked with DH, RSA, and other
structures.  Unfortunately, those accessors did not exist before the structures
were made opaque, so there has to be conditional code to handle the old and new
cases side by side.

While here, use OpenSSL AES-CTR support when present.

It is recommended that correct operation of the SSH mode be verified with
either version of OpenSSL.